refactor: Rename _serde.cjs -> serialization.cjs
