+++ scripts/eslint-plugin-custom-rules/cjs/esm_importer.worker.mjs
@@ -1,6 +1,6 @@
 /* eslint-env worker, node */
 import { parentPort } from "node:worker_threads";
-import { serialize } from "./_serde.cjs";
+import { serialize } from "./serialization.cjs";
 
 parentPort.on("message", async (id) => {
   parentPort.postMessage(serialize(await import(id)));
